manifest: change LEB- to staging- in all manifests
The LEB (Linaro Evaluation Build) designation is made after evaluating
the quality of a build by the release team. A less ambiguous term is
staging which is used to indicate that a given that a build has been
made with a kernel who's patches are not upstreamable at the time of
the build. This is in contrast to builds that do not have the staging-
(or LEB- prefix before this change) which are made with kernels that
only contain upstreamable components.

Signed-off-by: Zach Pfeffer <zach.pfeffer@linaro.org>