landing-snowball.xml: rename staging- to landing-
Linaro Landing Teams deliver kernels that may be based on Linaro's
kernel/linaro-android.git which is itself based on Linaro's common
kernel or they may deliver kernels that are not (based on Linus HEAD
or a stable kernel release). To distinguish these 2 cases staging- will
indicate those kernels that are based on Linaro's kernels and landing-
will be based on kernels that are not.

Change-Id: If13f767b5da00e93b3f1627c629e21909b4c3b56
Signed-off-by: Mathieu J. Poirier <mathieu.poirier@linaro.org>