manifests: Add tracking-vexpress-tests.xml
We are adding a copy of tracking-vexpress.xml with
sched tests which are hosted in the private git
repository. This is simply duplicate work but
for now we have to live with it to have the
tests in restricted builds. Once we have an overlay
mechanism to build and deploy the tests, we can
get rid of the duplicate manifest.

Change-Id: Ibafd58c814109f740c863b1b6ad49ca01a19b683
Signed-off-by: Vishal Bhoj <vishal.bhoj@linaro.org>